introduced neopixel skill and unsubscribing to topics
neopixel skill working
noise of PIR when enabling the neopixel module needs further evaluation, maybe setup driven?
introduced feature to unsubscribe topics when a skill is stopped, as this is not supported by the mqtt driver this is just removing the callback object from the mqtt subscription list in the user mqtt module.